refactor(figma-plugin): P2 — split pure helpers out of extract.ts → extract-pure.ts
Pure mechanical refactor — no behaviour change. Moves the leaf helpers
that operate only on type-level Plugin API shapes (SceneNode, RGBA,
Paint, GradientPaint, FrameNode axis enums) into a new
`src/extract-pure.ts`. Async Plugin API consumers (`extractScene`,
`walk`, `captureInstanceMetadata`) stay in extract.ts.

**Helpers moved (12 of them):**
- Color: `paintToColor`, `rgbaToCss`, `hex2`, `gradientToCss`,
  `gradientFallbackFlat`
- Type / layout mapping: `mapNodeType`, `mapPrimaryAxisAlign`,
  `mapCounterAxisAlign`, `mapAxisSize`
- Library recognition fallback: `audioWidgetKindFromName`
- Vector heuristic: `isPureVectorIllustration`
- Font catalogue: `collectFontFamilyAssets`

**Why:** extract.ts had grown to 928 lines mixing pure helpers with
async Plugin API surface. Three tsconfig entries now include both
files (`code.tsconfig.json`, `headless.tsconfig.json`); `ui.tsconfig.json`
stays as-is (UI doesn't import extract). The split makes Agent A's
Python REST port at `tools/import-design/figma_rest_export.py` easier
to keep in sync — the helpers it mirrors all live in one file now.

**Sizes:**
- extract.ts:   928 → 788 lines (−140, −15%)
- extract-pure.ts: 209 lines (new)
- headless.js bundle: 25 240 → 25 265 bytes (+25 bytes, ESM import
  overhead under esbuild minify; well under the 50 KB MCP cap)

**Verification (CLAUDE.md "tests ship with fixes"):**
- `npm run typecheck` (all three entries) — clean.
- `npm run build` — clean. Headless 25 265 bytes (24 735 headroom).
- `npm test` — 16/16 pass (no regression in existing TS suite).

The conformance gate shipped in P3-alt (`planning/fixtures/figma-plugin/
conformance/diff_envelopes.py`) will catch any silent behaviour drift
between the plugin's headless bundle and Agent A's Python REST port
on the next re-capture cycle.

Roadmap: P1 (headless MCP bridge, #3227) + P3-alt (envelope
conformance test, planning fixture + diff tool) shipped earlier; this
is P2 — the cleanest of the three to land last because P3-alt is now
in place to catch any drift. Full NodeProvider abstraction (the
"facade" half of the original P2) is deferred until there's a second
TS consumer that benefits from it.
